Fix pictures in example
I was hardcoding the fb CDN urls which do not have any guarantees of working in the future. Instead, we can use the graph endpoint that redirects to the proper cdn url and is guaranteed to work all the time. I also used github url for Jason as I couldn't find his fb account :p